fix sign warning

+++ gst_bridge/src/rosimagesrc.cpp
@@ -681,9 +681,9 @@ static void rosimagesrc_sub_cb(Rosimagesrc * src, sensor_msgs::msg::Image::Const
   {
     if(!(src->step == msg->step / msg->width))
       RCLCPP_ERROR(src->logger, "image format changed during playback, step %d != %d", src->step, msg->step/msg->width);
-    if(!(src->height == msg->height))
+    if(!(src->height == (int) msg->height))
       RCLCPP_ERROR(src->logger, "image format changed during playback, height %d != %d", src->height, msg->height);
-    if(!(src->width == msg->width))
+    if(!(src->width == (int) msg->width))
       RCLCPP_ERROR(src->logger, "image format changed during playback, width %d != %d", src->width, msg->width);
     if(!(src->endianness == (msg->is_bigendian ? G_BIG_ENDIAN : G_LITTLE_ENDIAN)))
       RCLCPP_ERROR(src->logger, "image format changed during playback, endianness %d != %d", src->endianness, (msg->is_bigendian ? G_BIG_ENDIAN : G_LITTLE_ENDIAN));
